+++ testapp/testapp/testmain/fixtures.py
@@ -9,14 +9,12 @@ from factory.django import DjangoModelFactory, FileField
 from django_afip import models
 
 
-def _get_key_file():
-    basepath = settings.BASE_DIR
-    return open(os.path.join(basepath, 'test.key'))
+def _key_file():
+    return open(os.path.join(settings.BASE_DIR, 'test.key'))
 
 
-def _get_cert_file():
-    basepath = settings.BASE_DIR
-    return open(os.path.join(basepath, 'test.crt'))
+def _cert_file():
+    return open(os.path.join(settings.BASE_DIR, 'test.crt'))
 
 
 class UserFactory(DjangoModelFactory):
@@ -76,8 +74,8 @@ class TaxPayerFactory(DjangoModelFactory):
     name = 'John Smith'
     cuit = 20329642330
     is_sandboxed = True
-    key = FileField(from_func=_get_key_file)
-    certificate = FileField(from_func=_get_cert_file)
+    key = FileField(from_func=_key_file)
+    certificate = FileField(from_func=_cert_file)
 
 
 class TaxPayerProfileFactory(DjangoModelFactory):